+++ src/main/__tests__/apiClientCordova.spec.ts
@@ -1,6 +1,6 @@
 import fetchMock from 'jest-fetch-mock'
 
-import { defineWindowProperty, headers } from './testHelpers'
+import { defineWindowProperty, headers, mockPkceValues, mockPkceWindow } from './testHelpers'
 import ApiClient from '../apiClient'
 import createEventManager from '../identityEventManager'
 import createUrlParser from '../urlParser'
@@ -17,7 +17,8 @@ function apiClientAndEventManager() {
       clientId,
       domain,
       language: 'en',
-      sso: false
+      sso: false,
+      pkceEnforced: false,
     },
     eventManager,
     urlParser: createUrlParser(eventManager)
@@ -30,6 +31,7 @@ beforeEach(() => {
 
   defineWindowProperty('location')
   defineWindowProperty('cordova', {})
+  defineWindowProperty('crypto', mockPkceWindow)
 
   delete window.handleOpenURL
   fetchMock.resetMocks()
@@ -607,7 +609,7 @@ describe('loginWithSocialProvider', () => {
           response_type: 'token',
           scope: 'openid profile email phone',
           display: 'page',
-          provider: 'facebook'
+          provider: 'facebook',
         }),
       '_system'
     )
@@ -744,7 +746,8 @@ describe('loginWithSocialProvider', () => {
           redirect_uri: redirectUri,
           scope: 'openid profile email phone',
           display: 'page',
-          provider: 'facebook'
+          provider: 'facebook',
+          ...mockPkceValues,
         }),
       '_system'
     )
@@ -779,7 +782,8 @@ describe('loginWithSocialProvider', () => {
           redirect_uri: redirectUri,
           scope: 'openid profile email phone',
           display: 'page',
-          provider: 'facebook'
+          provider: 'facebook',
+          ...mockPkceValues,
         }),
       '_blank'
     )