+++ src/main/main.ts
@@ -75,8 +75,8 @@ function checkParam<T>(data: T, key: keyof T) {
 }
 
 export function createClient(creationConfig: Config): Client {
-  checkParam(creationConfig, 'clientId')
   checkParam(creationConfig, 'domain')
+  checkParam(creationConfig, 'clientId')
 
   const { domain, clientId, language } = creationConfig