+++ src/main/pkceService.ts
@@ -22,8 +22,9 @@ function generateCodeVerifier(): string {
 
 function computeCodeChallenge(verifier: string): Promise<string> {
   const binaryChallenge = Buffer.from(verifier, 'utf-8')
-
-  return new Promise(resolve => {
-    window.crypto.subtle.digest('SHA-256', binaryChallenge).then(hash => resolve(encodeToBase64(hash)))
+    return new Promise(resolve => {
+    window.crypto.subtle.digest('SHA-256', binaryChallenge).then(hash => {
+      return resolve(encodeToBase64(hash))
+    })
   })
 }